+++ CHANGELOG.md
@@ -4,8 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.1.0] - 2021-10-28
+### Changed
 
+- `operator-rs` `0.3.0` → `0.4.0` ([#32]).
+- `stackable-hive-crd` `0.1.0` → `0.2.0` ([#32]).
+- `stackable-regorule-crd` `0.1.0` → `0.2.0` ([#32]).
+- `stackable-opa-crd` `0.4.1` → `0.5.0` ([#32]).
+- Adapted pod image and container command to docker image ([#32]).
+- Adapted documentation to represent new workflow with docker images ([#32]).
+
+[#32]: https://github.com/stackabletech/trino-operator/pull/32
+
+## [0.1.0] - 2021-10-28
 
 ### Changed
 - Switched to operator-rs tag 0.3.0 ([#21])